requirements.txt
streamlit
python.dotenv
langchain
pypdf
google-generativeai
